requirements.txt
streamlit
st-supabase-connection
streamlit_js_eval
pyyaml
streamlit-cookies-manager